Conflicts:
	openexchange-test/src/com/openexchange/ajax/chronos/bugs/ChronosBugsTestSuite.java
	openexchange-test/src/com/openexchange/ajax/task/TasksTest.java

+++ com.openexchange.message.timeline/src/com/openexchange/message/timeline/osgi/MessageTimelineEventHandler.java
@@ -80,7 +80,7 @@ public final class MessageTimelineEventHandler implements EventHandler {
                     Integer userId = (Integer) event.getProperty(SessiondEventConstants.PROP_USER_ID);
                     if (null != userId) {
                         // No active session left
-                        MessageTimelineManagement.getInstance().dropFor(userId, contextId);
+                        MessageTimelineManagement.getInstance().dropFor(userId.intValue(), contextId.intValue());
                     }
                 }
             } catch (final Exception e) {
